# openapi.json
{"openapi":"3.1.0","info":{"title":"API Prédiction RH & Sauvegarde","version":"0.1.0"},"paths":{"/":{"get":{"tags":["accueil"],"summary":"Accueil","description":"Sert `index.html` s'il est présent dans le dossier de travail, sinon renvoie 404.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/colonnes":{"get":{"tags":["prediction"],"summary":"Colonnes du modèle","description":"Liste les colonnes attendues par un modèle chargé, avec le seuil associé. Si le modèle est inconnu, renvoie les colonnes par défaut.","operationId":"get_colonnes_colonnes_get","parameters":[{"name":"modele","in":"query","required":true,"schema":{"type":"string","description":"Nom du modèle chargé, par exemple `top1`.","title":"Modele"},"description":"Nom du modèle chargé, par exemple `top1`."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/predict":{"post":{"tags":["prediction"],"summary":"Prédire et enregistrer","description":"Calcule la probabilité de départ, applique le seuil, puis enregistre le résultat. Une prédiction de 1 signifie un risque de départ.","operationId":"predict_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sauvegarder":{"post":{"tags":["resultats"],"summary":"Enregistrer une prédiction","description":"Enregistre une prédiction déjà calculée. Sans `employe_id`, les features sont copiées dans `details`.","operationId":"sauvegarder_prediction_sauvegarder_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SauvegardeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/resultats":{"get":{"tags":["resultats"],"summary":"Lister les prédictions","description":"Historique des prédictions, du plus récent au plus ancien, filtrable par nom.","operationId":"lister_resultats_resultats_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","description":"Nombre maximum de lignes renvoyées.","default":20,"title":"Limit"},"description":"Nombre maximum de lignes renvoyées."},{"name":"prenom","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filtre exact sur le prénom.","title":"Prenom"},"description":"Filtre exact sur le prénom."},{"name":"nom","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filtre exact sur le nom.","title":"Nom"},"description":"Filtre exact sur le nom."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/resultats/{resultat_id}":{"get":{"tags":["resultats"],"summary":"Détail d'une prédiction","description":"Renvoie la prédiction et ses features. Si `employe_id` est renseigné, les features viennent de `employes_features` ; sinon elles sont lues dans le JSON `details`.","operationId":"detail_resultat_resultats__resultat_id__get","parameters":[{"name":"resultat_id","in":"path","required":true,"schema":{"type":"integer","description":"Identifiant de la ligne dans `resultats`.","title":"Resultat Id"},"description":"Identifiant de la ligne dans `resultats`."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/annuaire":{"get":{"tags":["resultats"],"summary":"Annuaire des employés","description":"Liste les employés connus, avec l'indication s'ils ont quitté l'entreprise.","operationId":"lister_annuaire_annuaire_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Début de l'identifiant employé à rechercher.","title":"Q"},"description":"Début de l'identifiant employé à rechercher."},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","description":"Nombre maximum d'employés renvoyés.","default":2000,"title":"Limit"},"description":"Nombre maximum d'employés renvoyés."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/employes/{identifiant}":{"get":{"tags":["resultats"],"summary":"Fiche employé","description":"Renvoie les features d'un employé et son label réel, s'il existe.","operationId":"obtenir_employe_employes__identifiant__get","parameters":[{"name":"identifiant","in":"path","required":true,"schema":{"type":"integer","description":"Identifiant dans `employes_features`.","title":"Identifiant"},"description":"Identifiant dans `employes_features`."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/debug/modele/{modele}":{"get":{"tags":["debug"],"summary":"Inspecter un modèle","description":"Décrit le modèle chargé : type, scaler, colonnes, seuil et rôles des colonnes.","operationId":"debug_modele_debug_modele__modele__get","parameters":[{"name":"modele","in":"path","required":true,"schema":{"type":"string","description":"Nom du modèle chargé, par exemple `top1`.","title":"Modele"},"description":"Nom du modèle chargé, par exemple `top1`."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PredictionRequest":{"properties":{"modele":{"type":"string","title":"Modele","description":"Nom du modèle chargé, par exemple `top1`."},"features":{"additionalProperties":true,"type":"object","title":"Features","description":"Valeurs des colonnes du modèle."},"seuil":{"type":"number","title":"Seuil","description":"Seuil au-dessus duquel l'employé est classé à risque.","default":0.37},"prenom":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Prenom","description":"Prénom enregistré avec la prédiction.","default":"John"},"nom":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nom","description":"Nom enregistré avec la prédiction.","default":"Doe"},"employe_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Employe Id","description":"Identifiant annuaire. Vide si les features sont saisies à la main."}},"type":"object","required":["modele","features"],"title":"PredictionRequest","description":"Demande de prédiction. Le résultat est calculé puis enregistré."},"SauvegardeRequest":{"properties":{"prenom":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Prenom","description":"Prénom associé à la prédiction.","default":"John"},"nom":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nom","description":"Nom associé à la prédiction.","default":"Doe"},"modele_utilise":{"type":"string","title":"Modele Utilise","description":"Modèle qui a produit la prédiction."},"probabilite_de_quitter":{"type":"number","title":"Probabilite De Quitter","description":"Probabilité de départ, entre 0 et 1."},"prediction":{"type":"integer","title":"Prediction","description":"1 si le seuil est dépassé, sinon 0."},"libelle_prediction":{"type":"string","title":"Libelle Prediction","description":"Libellé lisible, par exemple Quitte ou Reste."},"seuil_applique":{"type":"number","title":"Seuil Applique","description":"Seuil utilisé pour trancher la prédiction."},"features":{"additionalProperties":true,"type":"object","title":"Features","description":"Features de la saisie. Ignorées en base si `employe_id` est renseigné."},"employe_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Employe Id","description":"Employé lié. Doit exister dans `employes_features` si renseigné."}},"type":"object","required":["modele_utilise","probabilite_de_quitter","prediction","libelle_prediction","seuil_applique"],"title":"SauvegardeRequest","description":"Prédiction déjà calculée, à enregistrer dans `resultats`."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}